updated styling for newTask.tsx
